% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customize_rstudio_docker.R
\name{customize_rstudio_docker}
\alias{customize_rstudio_docker}
\title{Customize rstudio docker container running in Azure Machine Learning Studio}
\usage{
customize_rstudio_docker(
  username,
  path_users = "~/cloudfiles/code/Users",
  git_setup = TRUE,
  git_name,
  git_email
)
}
\arguments{
\item{username}{User-specific username for the Azure Machine Learning Studio \
environment. This should match the user-specific directory nested directly
under \code{path_users}.}

\item{path_users}{The path for the Users directory on Azure Machine Learning
Studio.}

\item{use_git}{Whether or not to apply specific Git settings. If \code{TRUE},
copies user-specific \code{.gitconfig} file if found, installs git secrets, and
sets git credentials.}
}
\description{
Customize rstudio docker container running in Azure Machine Learning Studio
}
