% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mount-blob-storage.R
\name{unmount_blob_storage}
\alias{unmount_blob_storage}
\title{Unmount Azure Blob Storage}
\usage{
unmount_blob_storage(mount_point)
}
\arguments{
\item{mount_point}{The local directory where the container is mounted.}
}
\value{
Invisibly returns \code{TRUE} if the unmount was successful, otherwise \code{FALSE}.
}
\description{
Unmounts a previously mounted Azure Blob Storage container.
}
