% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project_aml.R
\name{create_project_aml}
\alias{create_project_aml}
\title{Create a project for Azure Machine Learning Studio}
\usage{
create_project_aml(
  path,
  use_renv = TRUE,
  use_python = TRUE,
  use_targets = TRUE,
  use_git = TRUE,
  use_github = FALSE,
  github_private = TRUE,
  r_version = NULL,
  py_version = NULL,
  dockerfile = TRUE
)
}
\arguments{
\item{path}{The path for the project directory. If it exists, it is used.
If it does not exist, it is created, provided that the parent path exists.
This value is passed the \code{path} argument of \code{usethis::create_project()}.}

\item{use_renv}{Whether or not to initialize the project to use the \link{renv}
(https://rstudio.github.io/renv/index.html) package. If \code{TRUE}, the
\code{renv.lock} file will be created within the project at \code{env/renv.lock} to
ensure any relevant Dockerfiles within the \code{env} directory have access.}

\item{use_python}{Whether or not to initialize \link{renv}
(https://rstudio.github.io/renv/index.html) python integration. This only
happens if \code{use_renv} is also \code{TRUE}.}

\item{use_targets}{Whether or not to initialize the project to use the
\href{https://docs.ropensci.org/targets/}{targets} package.}

\item{use_git}{Whether or not to initialize a Git repository. If \code{TRUE},
calls \code{usethis::use_git()}.}

\item{use_github}{Whether or not to connect the Git repository with GitHub.
If \code{TRUE} and \code{use_git} is also \code{TRUE}, calls \code{usethis::use_github()}.}

\item{github_private}{Whether or not the GitHub repository should be private.
This value is passed to the \code{private} argument of \code{usethis::use_github()}.
This argument is only evaluated if both \code{use_git} and \code{use_github} are \code{TRUE}.}

\item{r_version}{Declare a specific version of R to use for \link{renv}
(https://rstudio.github.io/renv/index.html) integration and/or Dockerfile
creation. If \code{NULL}, this will be set to the R version of the user's session.}

\item{py_version}{Declare a specific version of Python to be associated with
\href{https://rstudio.github.io/renv/index.html}{renv}. This argument is only
evaluated if both \code{use_renv} and \code{use_python} are \code{TRUE}.}

\item{dockerfile}{Whether or not to create a generic Dockerfile within the
project at \code{env/Dockerfile} for use when submitting R jobs to Azure Machine
Learning Studio compute clusters. If \code{use_rev} is \code{TRUE}, the Dockerfile will
include code to install \href{https://rstudio.github.io/renv/index.html}{renv} and
restore packages from the \code{renv.lock} file.}
}
\description{
Create a project for Azure Machine Learning Studio
}
