% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mount-blob-storage.R
\name{mount_blob_storage}
\alias{mount_blob_storage}
\title{Mount Azure Blob Storage using blobfuse2}
\usage{
mount_blob_storage(
  account_name,
  container_name,
  mount_point = fs::path("~/azureml-blobstore", container_name)
)
}
\arguments{
\item{account_name}{A character string specifying the Azure Storage account
name.}

\item{container_name}{A character string specifying the Azure Storage container
name.}

\item{mount_point}{A \code{fs_path} object or character string specifying the
local directory to mount the storage to. If not provided, defaults to
\verb{~/.azureml-blobstore/<container_name>}.}
}
\value{
Called for its side effects; returns invisibly. Prints a message
indicating whether the mount was successful.
}
\description{
Mounts an Azure Blob Storage container to a local directory using
\href{https://github.com/Azure/azure-storage-fuse}{blobfuse2}. If the storage
is already mounted, a message is displayed and the function exits early.
Requires that \code{blobfuse2} and the Azure CLI are installed and authenticated.
}
